gcf_gtas_validator/requirements.txt
functions-framework==3.*
flask==3.*
pandas==2.*
numpy==1.*
google-cloud-storage==2.*
google-auth==2.*